+++ src/renderer/src/voice/openai.ts
@@ -92,7 +92,7 @@ export class OpenAIVoiceProvider implements VoiceProvider {
 
           this.idleTimeout = setTimeout(() => {
             this.config?.onStopRecording?.();
-          }, 5000);
+          }, this.config?.idleTimeoutMs ?? 5000);
         }
       } catch (error) {
         // eslint-disable-next-line no-console